Moved main component to root level
