+++ src/app/query-explainer/components/commitment-data-section/commitment-data-section.component.html
