+++ src/app/query-explainer/components/input-data-section/input-data-section.component.scss
