+++ src/app/query-explainer/components/reproduced-result-section/reproduced-result-section.component.ts
