+++ src/app/query-explainer/components/verify-result-section/verify-result-section.component.scss
