+++ src/app/query-explainer/query-explainer-routing.module.ts
