+++ src/app/query-explainer/query-explainer.component.scss
