+++ src/app/query-explainer/query-explainer.mocks.ts
