+++ src/app/query-explainer/query-explainer.module.ts
@@ -3,17 +3,23 @@ import { CommonModule } from "@angular/common";
 import { QueryExplainerComponent } from "./query-explainer.component";
 import { MatIconModule } from "@angular/material/icon";
 import { AngularSvgIconModule } from "angular-svg-icon";
-import { DisplayHashModule } from "../display-hash/display-hash.module";
 import { RouterModule } from "@angular/router";
 import { HighlightModule } from "ngx-highlightjs";
-import { DynamicTableModule } from "../dynamic-table/dynamic-table.module";
 import { VerifyResultSectionComponent } from "./components/verify-result-section/verify-result-section.component";
 import { ReproducedResultSectionComponent } from "./components/reproduced-result-section/reproduced-result-section.component";
-import { InputDataSectionComponent } from './components/input-data-section/input-data-section.component';
-import { CommitmentDataSectionComponent } from './components/commitment-data-section/commitment-data-section.component';
+import { InputDataSectionComponent } from "./components/input-data-section/input-data-section.component";
+import { CommitmentDataSectionComponent } from "./components/commitment-data-section/commitment-data-section.component";
+import { DisplayHashModule } from "../components/display-hash/display-hash.module";
+import { DynamicTableModule } from "../components/dynamic-table/dynamic-table.module";
 
 @NgModule({
-    declarations: [QueryExplainerComponent, VerifyResultSectionComponent, ReproducedResultSectionComponent, InputDataSectionComponent, CommitmentDataSectionComponent],
+    declarations: [
+        QueryExplainerComponent,
+        VerifyResultSectionComponent,
+        ReproducedResultSectionComponent,
+        InputDataSectionComponent,
+        CommitmentDataSectionComponent,
+    ],
     imports: [
         CommonModule,
         MatIconModule,
