+++ src/app/query-explainer/query-explainer.service.ts
