+++ src/app/query-explainer/query-explainer.types.ts
